pyproject.toml
[project]
name = "chaosrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyMuPDF>=1.23.0",
    "tqdm>=4.66.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pillow>=12.1.0",
]
